pyproject.toml
[project]
name = "ai-meeting-saas"
version = "0.1.0"

description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.119.1",
    "flask>=3.1.2",
    "google-genai>=1.45.0",
    "jinja2>=3.1.6",
    "json-repair>=0.52.1",
    "passlib[bcrypt]>=1.7.4",
    "pyannote-audio==3.4.0",
    "python-dotenv>=1.1.1",
    "python-ffmpeg>=2.0.12",
    "python-jose[cryptography]>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "pywhispercpp>=1.3.3",
    "rich>=14.2.0",
    "torchaudio==2.8.0",
    "torchcodec>=0.8.0",
    "uvicorn>=0.38.0",
    "webvtt-py>=0.5.1",
    "whisperx>=3.7.4",
]
